+++ libs/alert/src/lib/alert.module.ts
@@ -3,10 +3,10 @@ import { CommonModule } from '@angular/common';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NzAlertModule } from 'ng-zorro-antd/alert';
 import { AlertComponent } from './alert/alert.component';
-import { IconModule, ICONS_TOKEN } from '@spryker/icon';
+import { IconModule, provideIcons, Icon } from '@spryker/icon';
 import errorIcon from './icons/error';
 
-const icons = [
+const icons: Icon[] = [
   {
     name: 'error',
     svg: errorIcon,
@@ -17,11 +17,6 @@ const icons = [
   imports: [CommonModule, IconModule, BrowserAnimationsModule, NzAlertModule],
   declarations: [AlertComponent],
   exports: [AlertComponent],
-  providers: [
-    {
-      provide: ICONS_TOKEN,
-      useValue: icons,
-    },
-  ],
+  providers: [provideIcons(icons)],
 })
 export class AlertModule {}