+++ libs/notification/src/lib/notification/icons/index.ts
@@ -1,0 +1,7 @@
+import errorIcon from './error';
+import successIcon from './success';
+import warningIcon from './warning';
+import infoIcon from './info';
+import removeIcon from './remove';
+
+export { errorIcon, successIcon, warningIcon, infoIcon, removeIcon };